Change input type to int

+++ challenge1/challenge1.go
@@ -1,6 +1,6 @@
 package challenge1
 
 
-func Run(n uint64) uint64  {
+func Run(n int) uint64  {
   return 0
 }
